# swagger.yaml
openapi: 3.0.3
info:
  title: Movie Mania 3.0
  description:
    MovieMania é o seu destino online para tudo relacionado a filmes! Nosso site permite que você crie e organize listas personalizadas de seus filmes favoritos, escreva comentários detalhados sobre os filmes que assistiu e fique por dentro dos lançamentos mais recentes. Explore uma vasta biblioteca de títulos, interaja com uma comunidade de cinéfilos e descubra novas recomendações baseadas nas suas preferências. Com uma interface amigável e recursos intuitivos, o MovieMania torna a experiência de assistir filmes ainda mais envolvente e divertida. Junte-se a nós e compartilhe sua paixão pelo cinema!
  version: 1.0.11
servers:
  - url: http://localhost:3333/
...
paths:
...
  /users/sign-up:
    post:
      tags:
        - users
      summary: 'Create user'
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUser'
      responses:
        '201':
          description: User Created Successfully
        '500':
          description: Internal Server Error.
  /users/profile:
    get:
      tags:
        - users
      summary: 'Get user profile'
      operationId: getUserProfile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '500':
          description: Internal Server Error.
  /users/login:
    post:
      tags:
        - users
      summary: 'Sign in user'
      operationId: login
      requestBody:
        description: Login user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginUser'
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JWT'
        '404':
          description: user does not exist
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/errorObject'
...
components:
  schemas:
    errorObject:
      type: object
      properties:
        description:
          type: string
          example: user not found
        code:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
    UserProfile:
      type: object
      properties:
        name:
          type: string
          example: "Itallo Vidal"
        favoriteGenres:
          type: array
          items:
            type: string
            example: "Ação"
...
    JWT:
      type: object
      properties:
        token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c
    LoginUser:
      type: object
      properties:
        password:
          type: string
          example: 1234567
        email:
          type: string
          example: Itallo@email.com
    CreateUser:
      type: object
      properties:
        username:
          type: string
          example: Itallo
        password:
          type: string
          example: 1234567
        email:
          type: string
          example: Itallo@email.com
        favoriteGenre1:
          type: number
          example: 12
        favoriteGenre2:
          type: number
          example: 28
...